pyproject.toml
[tool.poetry]
name = "ded"
version = "0.1.1"
description = "Helm dependency deduplication post-renderer"
authors = ["vduseev <vagiz.d@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/vduseev/ded"
repository = "https://github.com/vduseev/ded"
documentation = "https://github.com/vduseev/ded"
keywords = ["helm", "dependency", "post-renderer", "deduplication", "subchart"]

[tool.poetry.scripts]
ded = "ded:run"

[tool.poetry.dependencies]
python = "^3.6"
"ruamel.yaml" = "^0.16.10"

[tool.poetry.dev-dependencies]
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
